backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
supabase
pydantic-settings
python-multipart
